update python ci one iter
